Compare cached expires DateInstance as double, not via int64 cast
The expires getter's cache-hit check did
static_cast<int64_t>(dateInstance->internalNumber()) == impl.expires().
The cached DateInstance is returned to JS, so the user can mutate it via
setTime(NaN) (or any value TimeClip maps to NaN), after which the cast
is undefined per [conv.fpint]. Compare in the double domain instead;
NaN != x is well-defined and falls through to creating a fresh
DateInstance.

+++ src/jsc/bindings/webcore/JSCookie.cpp
@@ -730,7 +730,7 @@ JSC_DEFINE_CUSTOM_GETTER(jsCookiePrototypeGetter_expires, (JSGlobalObject * lexi
     if (impl.hasExpiry()) {
         if (thisObject->m_expires) {
             auto* dateInstance = thisObject->m_expires.get();
-            if (static_cast<int64_t>(dateInstance->internalNumber()) == impl.expires()) {
+            if (dateInstance->internalNumber() == static_cast<double>(impl.expires())) {
                 return JSValue::encode(dateInstance);
             }
         }
